requirements.txt
streamlit
PIL
pandas
seaborn
numpy
pandas
matplotlib.pyplot
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
typing
json
